+++ metric/humanparsing/mhp_extension/detectron2/detectron2/modeling/anchor_generator.py
@@ -43,10 +43,10 @@ class BufferList(nn.Module):
 def _create_grid_offsets(size: List[int], stride: int, offset: float, device: torch.device):
     grid_height, grid_width = size
     shifts_x = torch.arange(
-        offset * stride, grid_width * stride, step=stride, dtype=torch.float16, device=device
+        offset * stride, grid_width * stride, step=stride, dtype=torch.float32, device=device
     )
     shifts_y = torch.arange(
-        offset * stride, grid_height * stride, step=stride, dtype=torch.float16, device=device
+        offset * stride, grid_height * stride, step=stride, dtype=torch.float32, device=device
     )
 
     shift_y, shift_x = torch.meshgrid(shifts_y, shifts_x)
